third_party/blink/web_tests/wpt_internal/css/fit-width-text/text-grow-computed.html

<!DOCTYPE html>
<title>getComputedStyle().textFit</title>
<link rel="help" href="https://github.com/explainers-by-googlers/css-fit-text/blob/main/README.md">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/computed-testcommon.js"></script>
<body>
<div id="target" style="font-size:20px; line-height:1;"></div>
<script>
test_computed_value('text-fit', 'none');
test_computed_value('text-fit', 'grow');
test_computed_value('text-fit', 'shrink');

test_computed_value('text-fit', 'grow consistent', 'grow');
test_computed_value('text-fit', 'grow per-line');
test_computed_value('text-fit', 'grow per-line-all');

// TODO(crbug.com/417306102): The `limit` component should be a percentage value.
test_computed_value('text-fit', 'grow per-line 32px');
test_computed_value('text-fit', 'grow consistent 3em', 'grow 60px');

test_computed_value('text-fit', 'shrink per-line 4ch', 'shrink per-line 40px');
test_computed_value('text-fit', 'shrink consistent 2lh', 'shrink 40px');
</script>
</body>
